moves builder models up a directory
