+++ libraries/Pliant/Builders/AlterationModel.cs
@@ -1,6 +1,6 @@
 ﻿using System.Collections.Generic;
 
-namespace Pliant.Builders.Models
+namespace Pliant.Builders
 {
     public class AlterationModel
     {
