+++ libraries/Pliant/Builders/Expressions/RuleExpression.cs
@@ -1,4 +1,4 @@
-﻿using Pliant.Builders.Models;
+﻿using Pliant.Builders;
 using Pliant.Grammars;
 using System.Collections.Generic;
