+++ libraries/Pliant/Builders/Expressions/SymbolExpression.cs
@@ -1,4 +1,4 @@
-﻿using Pliant.Builders.Models;
+﻿using Pliant.Builders;
 using Pliant.Grammars;
 
 namespace Pliant.Builders.Expressions
