+++ libraries/Pliant/Builders/Fluent/FluentGrammarBuilder.cs
@@ -1,4 +1,4 @@
-﻿using Pliant.Builders.Models;
+﻿using Pliant.Builders;
 using Pliant.Grammars;
 using System;
 using System.Collections.Generic;
