+++ libraries/Pliant/Builders/SymbolModelType.cs
@@ -1,4 +1,4 @@
-﻿namespace Pliant.Builders.Models
+﻿namespace Pliant.Builders
 {
     public enum SymbolModelType
     {
