+++ libraries/Pliant/Pliant.csproj
@@ -49,15 +49,15 @@
     <Compile Include="Builders\Fluent\FluentGrammarBuilder.cs" />
     <Compile Include="Builders\Fluent\FluentProductionBuilder.cs" />
     <Compile Include="Builders\Fluent\FluentRuleBuilder.cs" />
-    <Compile Include="Builders\Models\AlterationModel.cs" />
-    <Compile Include="Builders\Models\GrammarModel.cs" />
-    <Compile Include="Builders\Models\NonTerminalModel.cs" />
-    <Compile Include="Builders\Models\ProductionModel.cs" />
-    <Compile Include="Builders\Models\LexerRuleModel.cs" />
-    <Compile Include="Builders\Models\ProductionReferenceModel.cs" />
-    <Compile Include="Builders\Models\ReachibilityMatrix.cs" />
-    <Compile Include="Builders\Models\SymbolModel.cs" />
-    <Compile Include="Builders\Models\SymbolModelType.cs" />
+    <Compile Include="Builders\AlterationModel.cs" />
+    <Compile Include="Builders\GrammarModel.cs" />
+    <Compile Include="Builders\NonTerminalModel.cs" />
+    <Compile Include="Builders\ProductionModel.cs" />
+    <Compile Include="Builders\LexerRuleModel.cs" />
+    <Compile Include="Builders\ProductionReferenceModel.cs" />
+    <Compile Include="Builders\ReachibilityMatrix.cs" />
+    <Compile Include="Builders\SymbolModel.cs" />
+    <Compile Include="Builders\SymbolModelType.cs" />
     <Compile Include="Forest\HorspoolSet.cs" />
     <Compile Include="Tree\ParseTreeEnumerator.cs" />
     <Compile Include="Forest\SinglePassForestNodeVisitorStateManager.cs" />
