+++ tests/Pliant.Tests.Unit/Builders/GrammarModelTests.cs
@@ -1,8 +1,8 @@
 ﻿using Microsoft.VisualStudio.TestTools.UnitTesting;
-using Pliant.Builders.Models;
+using Pliant.Builders;
 using Pliant.Grammars;
 
-namespace Pliant.Tests.Unit.Builders.Models
+namespace Pliant.Tests.Unit.Builders
 {
     [TestClass]
     public class GrammarModelTests
